RIA-7212 Refactor packages
